feat(home-page-tabs): add new tab to teachers home and move old beta home to storybook
