+++ src/components/HomePageTabs/CurriculumTab/CurriculumTab.stories.tsx
